+++ packages/brot/examples/09-training/main.ml
@@ -59,7 +59,10 @@ let () =
   show_trained "Word-level" wl_tok test_texts;
 
   (* Train Unigram: probabilistic subword segmentation *)
-  let uni_tok = train_unigram data ~vocab_size:100 ~show_progress:false in
+  let uni_tok =
+    train_unigram data ~vocab_size:100 ~show_progress:false
+      ~pre:(Pre_tokenizer.whitespace ())
+  in
   show_trained "Unigram" uni_tok test_texts;
 
   (* Training with special tokens *)